[FEAT/#16] core 모듈 안 util 패키징
