+++ core/src/main/java/com/terning/core/designsystem/component/button/RoundButton.kt
@@ -1,8 +1,10 @@
 package com.terning.core.designsystem.component.button
 
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import com.terning.core.designsystem.theme.TerningPointTheme
 
 @Composable
@@ -11,10 +13,11 @@ fun RoundButton(
     text: String = "",
     onClick: () -> Unit,
 ) {
-
+    TerningBasicButton(shape = RoundedCornerShape(8.dp),
+        onClick = onClick,
+        text = text)
 }
 
-// RoundedCornerShape(8.dp),
 @Preview(showBackground = true)
 @Composable
 fun RoundButtonPreview() {
