+++ core/src/main/java/com/terning/core/designsystem/component/button/TerningBasicButton.kt
@@ -4,8 +4,6 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.collectIsPressedAsState
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.material.ripple.LocalRippleTheme
-import androidx.compose.material.ripple.RippleAlpha
-import androidx.compose.material.ripple.RippleTheme
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Text
@@ -14,7 +12,6 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
 import androidx.compose.ui.tooling.preview.Preview
 import com.terning.core.designsystem.theme.TerningMain
@@ -22,6 +19,7 @@ import com.terning.core.designsystem.theme.TerningMain2
 import com.terning.core.designsystem.theme.TerningPointTheme
 import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.core.designsystem.theme.White
+import com.terning.core.util.NoRippleTheme
 
 @Composable
 fun TerningBasicButton(
@@ -37,7 +35,10 @@ fun TerningBasicButton(
         Button(
             modifier = Modifier.fillMaxWidth(),
             interactionSource = interactionSource,
-            colors = ButtonDefaults.buttonColors(containerColor = color, contentColor = White),
+            colors = ButtonDefaults.buttonColors(
+                containerColor = color,
+                contentColor = White
+            ),
             shape = shape,
             onClick = { onClick() }
         ) {
@@ -55,21 +56,9 @@ fun TerningBasicButtonPreview() {
     TerningPointTheme {
         TerningBasicButton(
             text = "Button",
-            shape = ButtonDefaults.shape,
+            shape = ButtonDefaults.textShape,
             onClick = {}
         )
     }
 }
 
-object NoRippleTheme : RippleTheme {
-    @Composable
-    override fun defaultColor() = Color.Unspecified
-
-    @Composable
-    override fun rippleAlpha(): RippleAlpha = RippleAlpha(
-        draggedAlpha = 0.0f,
-        focusedAlpha = 0.0f,
-        hoveredAlpha = 0.0f,
-        pressedAlpha = 0.0f
-    )
-}
