+++ core/src/main/java/com/terning/core/extension/Retrofit.kt
@@ -1,6 +1,5 @@
 package com.terning.core.extension
 
-
 fun String?.isJsonObject(): Boolean = this?.startsWith("{") == true && this.endsWith("}")
 
 fun String?.isJsonArray(): Boolean = this?.startsWith("[") == true && this.endsWith("]")
