+++ feature/src/main/java/com/terning/feature/onboarding/signin/SignInRoute.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -22,9 +21,8 @@ import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.flowWithLifecycle
 import androidx.navigation.NavHostController
-import com.terning.core.designsystem.component.button.TerningBasicButton
+import com.terning.core.designsystem.component.button.RoundButton
 import com.terning.core.designsystem.theme.TerningPointTheme
-import com.terning.core.designsystem.theme.TerningTheme
 import com.terning.core.extension.toast
 import com.terning.feature.R
 import com.terning.feature.home.navigation.navigateHome
@@ -77,7 +75,7 @@ fun SignInScreen(
             onSignInClick = { onSignInClick() },
             modifier = modifier.padding(horizontal = 20.dp)
         )
-        TerningBasicButton(text = "hi", shape = RoundedCornerShape(8.dp), onClick = onClick)
+        RoundButton(text= "hi", onClick = onClick)
     }
 }
